% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_choice_sorting.R
\name{assignment.choice}
\alias{assignment.choice}
\title{Assignment Choice Sorting}
\usage{
assignment.choice(survey.results, class.roster, varnames = c(),
  test.EIDs = c("speede", "sm56684", "tm29778"))
}
\arguments{
\item{survey.results}{A dataframe created from the qualtrics survey.}

\item{class.roster}{A dataframe of the class roster. Use a recently downloaded copy of the gradebook.}

\item{varnames}{A vector of character strings indicating the column names of the responses in the qualtrics survey.  Must be entered in the following order: eid, last name, first name, essay choice}

\item{test.EIDs}{A vector of character strings indicating the test student EIDs used in the class.  Defaults are for Speede, Student McStudent, and Tester McStudent. Function works even if these students are not present.}
}
\value{
Outputs three .csv files to a new folder in the working directory named to avoid any duplication or overwriting of previous files: "duplicated responses.csv" indicates any students who have submitted more than one unique response to the qualtrics survey. Ignores misspelled names and identical submissions; "missing responses.csv" indicates any students who did not submit a qualtrics survey response; "all responses.csv" contains all students including duplicates and missing students.
}
\description{
Use this function if you have an assignment in which students had a choice (e.g. essay topics) and indicated which assignment they chose through a qualtrics survey.
}
\examples{
assignment.choice(survey, roster, varnames=c("userid", "Q3", "Q5", "Q7"))
}

