% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradebook_clean.R
\name{gradebook_clean}
\alias{gradebook_clean}
\title{Gradebook Clean}
\usage{
gradebook_clean(gradebook, non_enrolled = c())
}
\arguments{
\item{gradebook}{A gradebook dataframe}

\item{non_enrolled}{A vector of EIDs of non-enrolled students}
}
\value{
A gradebook data frame with test and non-enrolled students removed, points possible line removed, periods removed from column names, and columns converted to numeric as appropriate
}
\description{
Clean up a gradebook all in one fell swoop
}
