% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maybe_as_numeric.R
\name{maybe_as_numeric}
\alias{maybe_as_numeric}
\title{Alex's maybe_numeric function}
\usage{
maybe_as_numeric(v)
}
\arguments{
\item{v}{A gradebook dataframe}
}
\description{
Internal function used to convert columns to numbers when they should be
}
