% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mute_check.R
\name{mute_check}
\alias{mute_check}
\title{Checking if any grades are muted}
\usage{
mute_check(gradebook)
}
\arguments{
\item{gradebook}{A gradebook file from Canvas}
}
\value{
Outputs True/False value
}
\description{
Internal use function to test whether any grades are muted in a given gradebook input
}

