% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_eid_add_grade.R
\name{Offline.grade}
\alias{Offline.grade}
\title{Generating Grades From Non-Canvas Grade Source}
\usage{
Offline.grade(qualtrics_vector, gradebook, assignment_col_number, point_value,
  missing_as_zero = TRUE, ignore_length = FALSE)
}
\arguments{
\item{qualtrics_vector}{A vector of EIDs for students who should be given full credit. Typically, from a qualtrics survey, though other sources are possible. Make sure that the vector does not contain any blank rows or all students will be given credit.}

\item{gradebook}{A gradebook file from Canvas. You do not need to worry about formatting.}

\item{assignment_col_number}{The column number of the assignment to give points to. If you are uncertain, use the names() function on your gradebook file to determine the correct value.}

\item{point_value}{The number of points that should be awarded for completing the assignment.}

\item{missing_as_zero}{Logical for whether students not included in the EID vector should be assigned a 0. Default is TRUE.}

\item{ignore_length}{Default is FALSE. If TRUE, then function will not check the length of your EIDs. Checking length is useful if students entered their own EIDs. Default minimum is < 5 will throw a warning. It is possible for a student to have a 4 character EID.}
}
\value{
Outputs an updated gradebook file. Note: R will convert () into "."s.  This will result in a file that Canvas will not accept. Copy/paste the single column as appropriate in a program outside of R and upload the same file that was downloaded. Also automatically outputs a warning message reporting which assignment you updated.
}
\description{
Use this function to add grades in a canvas gradebook for a complete/incomplete assignment based on a list of EIDs.
}

