% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attendance_random.R
\name{random_attendance}
\alias{random_attendance}
\title{Studio Attendance Randomization Function}
\usage{
random_attendance(piazza_survey, days = c("Monday", "Wednesday"),
  student_count = 25, ignore_list = c())
}
\arguments{
\item{piazza_survey}{A dataframe of the CSV file produced by the Piazza poll. Will not work with other data sources unless modified. See note in code if necessary.}

\item{days}{A listing of the days of the week for the course.  Default is Monday and Wednesday.}

\item{student_count}{The number of students you would like returned for each class day}

\item{ignore_list}{A vector of names that you want to exclude from the randomization}
}
\value{
Dataframe with three columns: Name, email address, and day of week they should be invited.
}
\description{
Use this function to randomize students to invite to the studio for lecture.
}
\details{
This function facilitates randomizing studio attendance and invititations for the online class. Create a poll in piazza asking students to complete it in order to add their name to the attendance drawing list.  (The poll can have a single option - "Yes, I'd like to attend")  Download the results of that poll and open the file in R.  Then process it using this function, which will place a CSV on your desktop with a list of 50 names, 25 for each day of the week, along with the student's email address.  This should facilitate contacting the students with directions for attending the studio.
}

