% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_point_calculator.R
\name{set.curve}
\alias{set.curve}
\title{Final Grade Calculator}
\usage{
\method{"McDS"}{For use with the McDaniel/Shaw version of Online310 only-
  calculation and rounding are non-standard!  It takes a given top.score value
  and calculates the number of students who would fall in each grade category.
  "Maximum points" return is top score in class if default used, otherwise
  returns inputted value.}(gradebook, method = c("McDS", "top points",
  "curve-a", "curve-ab"), top.score = "max score", curve.a = 10,
  curve.ab = 20, total.points = 500, grade.scale = c(93, 90, 87, 83, 80,
  77, 73, 70, 67, 63, 60))
}
\arguments{
\item{gradebook}{A dataframe of the final gradebook downloaded from Canvas.}

\item{method}{Four methods are supported: "McDS", "top points", "curve-a", and "curve-ab".  See Methods below for details.}

\item{top.score}{What value should be tested for McDS and top points methods. Default is maximum score in gradebook.}

\item{curve.a}{Percentage of students who should receive an A.  Must be input as an integer. For use with curve-a method only.}

\item{curve.ab}{Percentage of students who should receive either an A or a B. Must be input as an integer. For use with curve-ab method only.}

\item{total.points}{What is the total number of points in the class? For use with curve-a and curve-ab methods.}

\item{grade.scale}{A vector of length 11 indicating the LOWER point values for each grade break A through D-.  Use only integers- the function will take care of rounding. Breaks do not need to be any particular order. Default is UT standard division of grades.}
}
\value{
List of 2. Dataframe with four columns: grade, scale, number, and percent, which refer to letter grade, point value at lower end of range, number of students who will receive that grade, and the percent of students who will receive that grade, respectively. Second item in list is the maximum number of points, which is both use and method dependent.
}
\description{
This funciton has two primary purposes: first, to calculate a curve based on a desired percentage of students to receive either an A or an A or a B; second, to calculate the number of students who would receive a certain grade if top point value were adjusted.
}
\examples{
set.curve(gradebook.file, method=c("top points"), top.score="475")
set.curve(gradebook.file, method=c("curve-ab"), curve.ab=18, total.points=450)
}

