% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_student_search.R
\name{test.student}
\alias{test.student}
\title{A quick test student search function}
\usage{
test.student(EIDs = c("speede", "sm56684", "tm29778"), df)
}
\arguments{
\item{EIDs}{List of tester student EIDs.  Defaults for Speede, Tester McStudent, and Student McStudent}

\item{df}{A dataframe of a class roster.  Typically a recent copy of the Canvas gradeboook.}
}
\value{
A vector of the row numbers in which there are test students
}
\description{
This function identifies which rows in a data frame contain test students.  Required for other functions in package.  Automatically extracts line for universal "Student, Tester" as well.
}
\examples{
test.student.search.func()
}
