% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradebook_clean.R
\name{gradebook_clean}
\alias{gradebook_clean}
\title{Gradebook Clean}
\usage{
gradebook_clean(gradebook, non_enrolled = c())
}
\arguments{
\item{gradebook}{A gradebook dataframe}

\item{non_enrolled}{A vector of EIDs of non-enrolled students}
}
\description{
Clean up a gradebook all in one fell swoop
}

